requirements.txt
torch==2.0.1
torchaudio==2.0.2
librosa==0.10.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
pydub==0.25.1
scipy==1.10.1
plotly==5.14.1